fix: #810 (use more vue-ish way for attaching listeners)
